add logo for groups
